Fix crash in tagger when next() is called before parse()
Adds a null pointer check to the agenda_ member in next before accessing it.
Normally, agenda_ is allocated when parse is called, so the call to next
will only seg fault if it is called before parse. For cases of empty input
from a user, parse may not get called, but next should return false in this
case without crashing.